fix deprecation warning on Scala 2.13.2

+++ src/test/scala/net/tisue/euler/94.scala
@@ -9,7 +9,7 @@ package net.tisue.euler
 
 object Problem94 extends Problem(94, "518408346") {
   def isSquare(n: Long) = {
-    val r = math.sqrt(n).toLong
+    val r = math.sqrt(n.toDouble).toLong
     r * r == n
   }
   def solve = {
